Added reset button| migrated controller
Added a functional reset button that stops the timer when clicked (so that the timer doesn't just restart from 0), but this required migrating the stop timer feature from ClockController to GameClock so ClockController was removed and the functions transferred with their corresponding buttons and variables
