+++ src/components/button/Button.js
@@ -2,6 +2,6 @@ import React from "react";
 import * as css from "./css.module.scss";
 
 const Button = (props) => {
-  return <button className={css.circular}>{props.children}</button>;
+  return <button className={css.circular} onClick = {props.onClick}>{props.children}</button>;
 };
 export default Button;
